+++ roles/oneagent/tests/ansible/util.py
@@ -1,4 +1,4 @@
-from technology.constants import ERROR_MESSAGES_FILE
+from ansible.constants import ERROR_MESSAGES_FILE
 from util.common_utils import read_yaml_file
